requirements.txt
opencv_python==3.4.0.12
scipy==0.19.1
numpy==1.13.1+mkl
Pillow==5.3.0
python_dateutil==2.7.3
pytesseract==0.2.5
